requirements.txt
streamlit
yfinance
pandas
pandas_ta
google-generativeai
feedparser
plotly